Move  state logic to a new component
